+++ docs/start/changelog.rst
@@ -3,6 +3,11 @@
 ChangeLog
 =========
 
+0.13.3 - TBD
+   * Fixed a bug with giving an array of complex messgaes to target.script where
+     it would send the messages to all devices rather than just the devices you
+     care about.
+
 0.13.2 - 7 April 2019
    * Fixed behaviour when you provide a list of complex messages to run_with
    * Made HardCodedSerials more efficient when the afr has already found devices